Harmonized Artnet.h
Changes most defines to const, documentation showed that this is more efficient.
Added the dchp boolean which indicated if the connection was made with dchp. In case it does the maintain function need frequent calling.
Not tested, still have to verify weither this works with the udp library.
Additionally I enclosed the Art-Net 4 protocol specification to reference to.