Bugfix in ArtPollReply
Due to faulty condition the poll was send 4xMAX_NUM_UNIVERSES.
Which off course was incorrect.
Also it made no sense to implement a retry because there is no way to know the controller received the message.
Currently most field of the ArtPollReply are empty. I added a loadDefaults() function to insert it into the begin() function. That way the struct is loaded with all necessary details.