Added DCHP maintain & loadDefaults
Addition of these two function makes sure all default settings are loaded during startup.
maintain DCHP was added to the read() function, every time the library was read it will now check if nothing was changed on DCHP. This only when a DCHP Ip was assinged at begin. In case of static IP this function is ignored.